* Moved start trip code. * Updated Location Manager code.
