+++ KT20/SplashViewController.swift
@@ -21,10 +21,10 @@ class SplashViewController: UIViewController {
         super.viewWillAppear(animated)
         self.navigationController?.navigationBar.isHidden = true
         
-        LocationManager.shared.getUserLocation(location: { (location) in
-            print(location?.coordinate.latitude ?? 0.0)
-            print(location?.coordinate.longitude ?? 0.0)
-        })
+//        LocationManager.shared.getUserLocation(location: { (location) in
+//            print(location?.coordinate.latitude ?? 0.0)
+//            print(location?.coordinate.longitude ?? 0.0)
+//        })
     }
     
     func playAnimation(){
